convert to more Pythonic access of dicts
Improving readability, fewer lines. Also used the [py3.8 walrus](https://docs.python.org/3/whatsnew/3.8.html#assignment-expressions) since this action is on 3.8 anyhow.

Technically this changes logic, but it should be an improvement. I mean, "if FOO in dict" will be true even if the value of the entry is None or is ''. So this switches to a stringy truthiness.